# Optimize writing repeated columns