--- a/level_amd64.go
+++ b/level_amd64.go
@@ -2,5 +2,17 @@
 
 package parquet
 
+import "golang.org/x/sys/cpu"
+
 //go:noescape
-func memset(dst []byte, src byte)
+func memsetAVX2(dst []byte, src byte)
+
+func memset(dst []byte, src byte) {
+	if len(dst) >= minLenAVX2 && cpu.X86.HasAVX2 {
+		memsetAVX2(dst, src)
+	} else {
+		for i := range dst {
+			dst[i] = src
+		}
+	}
+}