I added a benchmark for repeated columns which highlighted that in cases where we are working on a small number of values (e.g. a `[]int32` struct field with a couple elements) the overhead of doing calls to the assembly functions would exceed the simple Go version of the code.

To address this issue, I modified the code to inline the check of whether applying an optimization is useful based on the number of rows we are working on, and otherwise fallback to the regular Go code when there are too few items.

Since we were now performing the check ahead of calling the assembly routines, I also moved the check of CPU features out of the function so we don't have to ensure that AVX2 is supported in the assembly code.

```
name                                      old time/op  new time/op  delta
GenericBuffer/repeatedInt32Column/go1.18  47.7µs ±14%  43.4µs ± 1%  -9.16%  (p=0.000 n=10+8)
GenericWriter/repeatedInt32Column/go1.18  54.6µs ± 0%  53.3µs ± 0%  -2.24%  (p=0.000 n=8+9)

name                                      old row/s    new row/s    delta
GenericBuffer/repeatedInt32Column/go1.18   21.1M ±13%   23.1M ± 1%  +9.37%  (p=0.000 n=10+8)
GenericWriter/repeatedInt32Column/go1.18   18.3M ± 0%   18.7M ± 0%  +2.29%  (p=0.000 n=8+9)
```